+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - The ability for 3rd party addons to fetch the ID of the post being currently indexed.
 
+### Changed
+- `post_author` field changed from Numeric to Text. *REQUIRES INDEX RECREATION AND REINDEXING*
+
 ### Fixed
 - `author` parameter queries.